+++ ChangeLog.md
@@ -1,9 +1,13 @@
 ## 3.1.0
+Features:
 * Add 25 and 50 FPS speed options.
 * Add buttons to clear keyboard shortcuts.
 * Add a file browser option for showing hidden files.
 * Add view position and zoom controls to the view tool, and a view zoom control to the tool bar.
 * Make the timeline ticks more visible.
+* Add a keyboard shortcut (Ctrl+G) for enabling the grid.
+* Add a keyboard shortcut (Ctrl+N) for toggling whether OCIO is enabled.
+* Add a keyboard shortcut (Ctrl+K) for toggling whether the LUT is enabled.
+Fixes:
 * Fix for image sequences with uppercase file extensions.
 * Fixes for building with BMD support.
-* Add a keyboard shortcut (Ctrl+G) for enabling the grid.